:recycle: Move tests to separate file
